+++ src/main/java/za/co/hpsc/web/models/ipsc/request/MatchRequest.java
@@ -1,0 +1,20 @@
+package za.co.hpsc.web.models.ipsc.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.time.LocalDate;
+import java.util.List;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class MatchRequest {
+    private LocalDate matchDate;
+    private String matchName;
+    private String club;
+    private List<MatchStageRequest> stages;
+}