+++ src/main/java/za/co/hpsc/web/models/ipsc/request/MatchStageRequest.java
@@ -1,0 +1,16 @@
+package za.co.hpsc.web.models.ipsc.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class MatchStageRequest {
+    private Long matchId;
+    private Integer stageNumber;
+    private String stageName;
+}